refactor(provider): split oversized tests.rs and types.rs (#6285)
