+++ src/core/agent_task_provider/tests/mod.rs
@@ -1,0 +1,16 @@
+use super::*;
+use crate::core::agent_task::{
+    AgentTaskExecutor, AgentTaskLimits, AgentTaskPolicy, AgentTaskWorkspace,
+    AgentTaskWorkspaceMode, AgentToolExecutionLocation, AgentToolPolicyRule,
+};
+use crate::core::agent_task_scheduler::{
+    AgentTaskCancellationToken, AgentTaskExecutionContext, AgentTaskPlan, AgentTaskScheduler,
+};
+use std::fs;
+
+mod common;
+mod manifest_tests;
+mod outcome_tests;
+mod scheduler_tests;
+mod selection_tests;
+mod workspace_secrets_tests;
