+++ src/core/agent_task_provider/types.rs
@@ -89,225 +89,6 @@ pub struct AgentTaskExecutorProvider {
     pub extra: BTreeMap<String, Value>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeContract {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub capabilities: Vec<String>,
-    #[serde(
-        default,
-        skip_serializing_if = "AgentTaskRuntimeLifecycleStates::is_empty"
-    )]
-    pub lifecycle_states: AgentTaskRuntimeLifecycleStates,
-    #[serde(
-        default,
-        skip_serializing_if = "AgentTaskRuntimeNormalization::is_empty"
-    )]
-    pub normalization: AgentTaskRuntimeNormalization,
-    #[serde(default, skip_serializing_if = "AgentTaskRuntimeApplyBack::is_empty")]
-    pub apply_back: AgentTaskRuntimeApplyBack,
-    /// Runtime dependency reconciliation contract: packages the runtime/overlay
-    /// (e.g. "WordPress 7.0 supplies this Composer package") provides, which a
-    /// staged provider plugin must NOT vendor/shadow. Homeboy validates the
-    /// effective staged plugin against this before dispatch so a stale vendored
-    /// runtime library fails with an actionable owner/contract message instead of
-    /// a raw PHP fatal during plugin activation (#6223).
-    #[serde(
-        default,
-        skip_serializing_if = "AgentTaskRuntimeStagingContract::is_empty"
-    )]
-    pub staging: AgentTaskRuntimeStagingContract,
-}
-
-impl AgentTaskRuntimeContract {
-    fn is_empty(&self) -> bool {
-        self.capabilities.is_empty()
-            && self.lifecycle_states.is_empty()
-            && self.normalization.is_empty()
-            && self.apply_back.is_empty()
-            && self.staging.is_empty()
-    }
-}
-
-/// Declares the runtime dependency reconciliation surface for a provider's
-/// staged plugins. Core is runtime-agnostic: it does not know that a package is
-/// a Composer dependency or that the owner is WordPress — the declaring
-/// extension supplies the package name, the owner that provides it at runtime,
-/// and the vendor subpaths a staged plugin would use to shadow it. Homeboy uses
-/// this to reconcile staged plugins before dispatch (#6223).
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeStagingContract {
-    /// Packages the runtime/overlay owns. A staged provider plugin that vendors
-    /// any of these would shadow the runtime-provided version and is refused
-    /// before dispatch with an actionable owner/package/contract error.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub reconciled_packages: Vec<AgentTaskRuntimeReconciledPackage>,
-    /// Optional human-facing remediation appended to reconciliation conflict
-    /// errors (e.g. how to rebuild the staged plugin without the vendored copy).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub remediation: Option<String>,
-    /// When true (default), Homeboy validates staged plugins against this
-    /// contract before dispatch. Set false to declare the contract for evidence
-    /// without enforcing the pre-dispatch gate (e.g. when WP Codebox owns the
-    /// authoritative readiness check and returns a structured result).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub validate_before_dispatch: Option<bool>,
-    #[serde(flatten, default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub extra: BTreeMap<String, Value>,
-}
-
-impl AgentTaskRuntimeStagingContract {
-    pub fn is_empty(&self) -> bool {
-        self.reconciled_packages.is_empty()
-            && self.remediation.is_none()
-            && self.validate_before_dispatch.is_none()
-            && self.extra.is_empty()
-    }
-
-    /// Whether the pre-dispatch validation gate is enforced. Defaults to true
-    /// when any reconciled package is declared, so declaring a package is enough
-    /// to opt into the gate; an explicit `validate_before_dispatch: false`
-    /// records the contract for evidence/Codebox delegation without gating.
-    pub fn enforces_pre_dispatch(&self) -> bool {
-        self.validate_before_dispatch.unwrap_or(true) && !self.reconciled_packages.is_empty()
-    }
-}
-
-/// A single package the runtime owns that staged provider plugins must not
-/// vendor. Generic by design: `name` is the package identity, `owner` is the
-/// runtime/overlay that supplies it (named in conflict errors), and
-/// `vendor_subpaths` are the staged-plugin-relative paths whose presence proves
-/// the staged plugin shadows the runtime copy (e.g. `vendor/<org>/<pkg>`).
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeReconciledPackage {
-    /// Package identity (e.g. a Composer package name `org/library`).
-    pub name: String,
-    /// The runtime/overlay that supplies this package at runtime. Surfaced in
-    /// conflict errors so the operator knows who owns the canonical copy.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub owner: Option<String>,
-    /// Staged-plugin-relative paths that, if present in the effective staged
-    /// plugin, prove it vendors (and would shadow) the runtime-owned package.
-    /// When omitted, `vendor/<name>` is checked by default.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub vendor_subpaths: Vec<String>,
-    /// Optional reason/why this package is reconciled, surfaced in diagnostics.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub reason: Option<String>,
-    /// Optional per-package remediation, surfaced when this package conflicts.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub remediation: Option<String>,
-    #[serde(flatten, default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub extra: BTreeMap<String, Value>,
-}
-
-impl AgentTaskRuntimeReconciledPackage {
-    /// The staged-plugin-relative subpaths whose presence indicates a shadowed
-    /// runtime package. Falls back to `vendor/<name>` when none are declared.
-    pub fn effective_vendor_subpaths(&self) -> Vec<String> {
-        let declared: Vec<String> = self
-            .vendor_subpaths
-            .iter()
-            .map(|subpath| subpath.trim().trim_matches('/').to_string())
-            .filter(|subpath| !subpath.is_empty())
-            .collect();
-        if !declared.is_empty() {
-            return declared;
-        }
-        let name = self.name.trim().trim_matches('/');
-        if name.is_empty() {
-            return Vec::new();
-        }
-        vec![format!("vendor/{name}")]
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeApplyBack {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub mutation_artifacts: Vec<AgentTaskRuntimeMutationArtifact>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub requires_git_checkout: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub strategy: Option<String>,
-}
-
-impl AgentTaskRuntimeApplyBack {
-    fn is_empty(&self) -> bool {
-        self.mutation_artifacts.is_empty()
-            && self.requires_git_checkout.is_none()
-            && self.strategy.is_none()
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeMutationArtifact {
-    pub name: String,
-    pub path: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub kind: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub semantic_key: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub apply_method: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeLifecycleStates {
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub execution_states: BTreeMap<String, AgentTaskExecutionState>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub outcome_statuses: BTreeMap<String, AgentTaskOutcomeStatus>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub failure_classifications: BTreeMap<String, AgentTaskFailureClassification>,
-}
-
-impl AgentTaskRuntimeLifecycleStates {
-    fn is_empty(&self) -> bool {
-        self.execution_states.is_empty()
-            && self.outcome_statuses.is_empty()
-            && self.failure_classifications.is_empty()
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeNormalization {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub status_path: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub summary_path: Option<String>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub output_artifacts: Vec<AgentTaskRuntimeOutputArtifactMapping>,
-}
-
-impl AgentTaskRuntimeNormalization {
-    fn is_empty(&self) -> bool {
-        self.status_path.is_none()
-            && self.summary_path.is_none()
-            && self.output_artifacts.is_empty()
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskRuntimeOutputArtifactMapping {
-    pub name: String,
-    #[serde(
-        default,
-        rename = "type",
-        alias = "artifact_type",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub artifact_type: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub artifact_schema: Option<String>,
-    pub path: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub kind: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub mime: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-}
-
 fn default_provider_schema() -> String {
     AGENT_TASK_EXECUTOR_PROVIDER_SCHEMA.to_string()
 }
@@ -318,49 +99,6 @@ fn default_request_schema() -> String {
 
 fn default_outcome_schema() -> String {
     AGENT_TASK_OUTCOME_SCHEMA.to_string()
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskProviderSecretRequirement {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_string_or_vec")]
-    pub env: Vec<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub required: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub purpose: Option<String>,
-    #[serde(flatten)]
-    pub extra: BTreeMap<String, Value>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskProviderSecretEnvRequirement {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub schema: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_string_or_vec")]
-    pub env: Vec<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub when: Option<Value>,
-    #[serde(flatten)]
-    pub extra: BTreeMap<String, Value>,
-}
-
-fn deserialize_string_or_vec<'de, D>(deserializer: D) -> Result<Vec<String>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let value = Option::<Value>::deserialize(deserializer)?;
-    Ok(match value {
-        Some(Value::String(item)) => vec![item],
-        Some(Value::Array(items)) => items
-            .into_iter()
-            .filter_map(|item| item.as_str().map(str::to_string))
-            .collect(),
-        _ => Vec::new(),
-    })
 }
 
 fn is_false(value: &bool) -> bool {
@@ -542,11 +280,11 @@ pub struct AgentTaskProviderArtifactPattern {
     pub extra: BTreeMap<String, Value>,
 }
 
-fn default_metadata() -> Value {
+pub(super) fn default_metadata() -> Value {
     Value::Object(Default::default())
 }
 
-fn is_empty_metadata(value: &Value) -> bool {
+pub(super) fn is_empty_metadata(value: &Value) -> bool {
     value.as_object().is_none_or(|object| object.is_empty())
 }
 
@@ -637,184 +375,4 @@ pub(crate) fn wildcard_match(pattern: &str, value: &str) -> bool {
         remainder = &remainder[index + part.len()..];
     }
     !anchored_end || value.ends_with(parts[parts.len() - 1])
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct AgentTaskProviderWorkspaceMaterialization {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cwd: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub requires_git: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub write_scope: Option<String>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub artifact_paths: Vec<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub spec: Option<WorkspaceMaterializationSpec>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub mounts: Vec<WorkspaceMountSpec>,
-    #[serde(default, skip_serializing_if = "AgentTaskRuntimeApplyBack::is_empty")]
-    pub apply_back: AgentTaskRuntimeApplyBack,
-    #[serde(flatten, default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub extra: BTreeMap<String, Value>,
-}
-
-impl AgentTaskProviderWorkspaceMaterialization {
-    pub fn requires_cwd_git_checkout(&self) -> bool {
-        self.apply_back.requires_git_checkout == Some(true)
-            || self.requires_git == Some(true)
-            || self.cwd.as_deref() == Some("git_checkout")
-    }
-
-    pub fn validate(&self) -> Result<(), Vec<String>> {
-        let mut errors = Vec::new();
-        if let Some(spec) = &self.spec {
-            errors.extend(
-                spec.validation_errors()
-                    .into_iter()
-                    .map(|error| format!("spec.{error}")),
-            );
-        }
-        for (index, mount) in self.mounts.iter().enumerate() {
-            errors.extend(
-                mount
-                    .validation_errors()
-                    .into_iter()
-                    .map(|error| format!("mounts[{index}].{error}")),
-            );
-        }
-        if errors.is_empty() {
-            Ok(())
-        } else {
-            Err(errors)
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct WorkspaceMaterializationSpec {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub handle: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub repo: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub host_path: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub target_path: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub mode: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub materialization: Option<String>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub mounts: Vec<WorkspaceMountSpec>,
-    #[serde(
-        default = "default_metadata",
-        skip_serializing_if = "is_empty_metadata"
-    )]
-    pub metadata: Value,
-    #[serde(flatten, default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub extra: BTreeMap<String, Value>,
-}
-
-impl WorkspaceMaterializationSpec {
-    pub fn validate(&self) -> Result<(), Vec<String>> {
-        let errors = self.validation_errors();
-        if errors.is_empty() {
-            Ok(())
-        } else {
-            Err(errors)
-        }
-    }
-
-    fn validation_errors(&self) -> Vec<String> {
-        let mut errors = workspace_mount_like_validation_errors(
-            self.handle.as_deref(),
-            self.repo.as_deref(),
-            self.host_path.as_deref(),
-            self.target_path.as_deref(),
-            self.mode.as_deref(),
-            self.materialization.as_deref(),
-        );
-        for (index, mount) in self.mounts.iter().enumerate() {
-            errors.extend(
-                mount
-                    .validation_errors()
-                    .into_iter()
-                    .map(|error| format!("mounts[{index}].{error}")),
-            );
-        }
-        errors
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct WorkspaceMountSpec {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub handle: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub repo: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub host_path: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub target_path: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub mode: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub materialization: Option<String>,
-    #[serde(
-        default = "default_metadata",
-        skip_serializing_if = "is_empty_metadata"
-    )]
-    pub metadata: Value,
-    #[serde(flatten, default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub extra: BTreeMap<String, Value>,
-}
-
-impl WorkspaceMountSpec {
-    pub fn validate(&self) -> Result<(), Vec<String>> {
-        let errors = self.validation_errors();
-        if errors.is_empty() {
-            Ok(())
-        } else {
-            Err(errors)
-        }
-    }
-
-    fn validation_errors(&self) -> Vec<String> {
-        workspace_mount_like_validation_errors(
-            self.handle.as_deref(),
-            self.repo.as_deref(),
-            self.host_path.as_deref(),
-            self.target_path.as_deref(),
-            self.mode.as_deref(),
-            self.materialization.as_deref(),
-        )
-    }
-}
-
-fn workspace_mount_like_validation_errors(
-    handle: Option<&str>,
-    repo: Option<&str>,
-    host_path: Option<&str>,
-    target_path: Option<&str>,
-    mode: Option<&str>,
-    materialization: Option<&str>,
-) -> Vec<String> {
-    let mut errors = Vec::new();
-    validate_non_blank_optional("handle", handle, &mut errors);
-    validate_non_blank_optional("repo", repo, &mut errors);
-    validate_non_blank_optional("host_path", host_path, &mut errors);
-    validate_non_blank_optional("target_path", target_path, &mut errors);
-    validate_non_blank_optional("mode", mode, &mut errors);
-    validate_non_blank_optional("materialization", materialization, &mut errors);
-    if host_path.is_some() && target_path.is_none() {
-        errors.push("target_path is required when host_path is set".to_string());
-    }
-    errors
-}
-
-fn validate_non_blank_optional(field: &str, value: Option<&str>, errors: &mut Vec<String>) {
-    if value.is_some_and(|value| value.trim().is_empty()) {
-        errors.push(format!("{field} must not be blank"));
-    }
 }
